+++ core/src/sql/date/mod.rs
@@ -3,19 +3,27 @@ mod timestamp;
 mod timestamptz;
 
 pub use interval::*;
+pub use timestamp::*;
+pub use timestamptz::*;
 
-fn base_date() -> crate::Result<chrono::NaiveDate> {
+fn base_date() -> crate::Result<Date> {
     base_datetime().map(|x| x.date())
 }
 
-fn base_datetime() -> crate::Result<chrono::NaiveDateTime> {
+fn base_datetime() -> crate::Result<Timestamp> {
     chrono::NaiveDate::from_ymd_opt(2000, 1, 1)
         .and_then(|x| x.and_hms_opt(0, 0, 0))
         .ok_or_else(|| crate::Error::Chrono("Invalid base date".to_string()))
 }
 
+/**
+ * Rust type for [date](https://www.postgresql.org/docs/current/datatype-datetime.html).
+ */
 #[cfg_attr(docsrs, doc(cfg(feature = "date")))]
-impl crate::ToSql for chrono::NaiveDate {
+pub type Date = chrono::NaiveDate;
+
+#[cfg_attr(docsrs, doc(cfg(feature = "date")))]
+impl crate::ToSql for Date {
     fn ty(&self) -> crate::pq::Type {
         crate::pq::types::DATE
     }
@@ -41,12 +49,12 @@ impl crate::ToSql for chrono::NaiveDate {
 }
 
 #[cfg_attr(docsrs, doc(cfg(feature = "date")))]
-impl crate::FromSql for chrono::NaiveDate {
+impl crate::FromSql for Date {
     /*
      * https://github.com/postgres/postgres/blob/REL_12_0/src/backend/utils/adt/date.c#L114
      */
     fn from_text(ty: &crate::pq::Type, raw: Option<&str>) -> crate::Result<Self> {
-        match chrono::NaiveDate::parse_from_str(crate::from_sql::not_null(raw)?, "%F") {
+        match Date::parse_from_str(crate::from_sql::not_null(raw)?, "%F") {
             Ok(date) => Ok(date),
             _ => Err(Self::error(ty, raw)),
         }
@@ -63,25 +71,25 @@ impl crate::FromSql for chrono::NaiveDate {
 }
 
 #[cfg_attr(docsrs, doc(cfg(feature = "date")))]
-impl crate::entity::Simple for chrono::NaiveDate {}
+impl crate::entity::Simple for Date {}
 
 #[cfg(test)]
 mod test {
     crate::sql_test!(
         date,
-        chrono::NaiveDate,
+        crate::Date,
         [
             (
                 "'1970-01-01'",
-                chrono::NaiveDate::from_ymd_opt(1970, 1, 1).unwrap()
+                crate::Date::from_ymd_opt(1970, 1, 1).unwrap()
             ),
             (
                 "'2010-01-01'",
-                chrono::NaiveDate::from_ymd_opt(2010, 1, 1).unwrap()
+                crate::Date::from_ymd_opt(2010, 1, 1).unwrap()
             ),
             (
                 "'2100-12-30'",
-                chrono::NaiveDate::from_ymd_opt(2100, 12, 30).unwrap()
+                crate::Date::from_ymd_opt(2100, 12, 30).unwrap()
             ),
         ]
     );