+++ core/src/sql/json.rs
@@ -1,3 +1,9 @@
+/**
+ * Rust type for [json](https://www.postgresql.org/docs/16/datatype-json.html).
+ */
+#[cfg_attr(docsrs, doc(cfg(feature = "json")))]
+pub type Json = serde_json::Value;
+
 #[cfg_attr(docsrs, doc(cfg(feature = "json")))]
 impl crate::ToSql for serde_json::Value {
     fn ty(&self) -> crate::pq::Type {