+++ core/src/sql/net/macaddr8.rs
@@ -1,5 +1,11 @@
+/**
+ * Rust type for [macaddr8](https://www.postgresql.org/docs/current/datatype-net-types.html).
+ */
 #[cfg_attr(docsrs, doc(cfg(feature = "net")))]
-impl crate::ToSql for macaddr::MacAddr8 {
+pub type MacAddr8 = macaddr::MacAddr8;
+
+#[cfg_attr(docsrs, doc(cfg(feature = "net")))]
+impl crate::ToSql for MacAddr8 {
     fn ty(&self) -> crate::pq::Type {
         crate::pq::types::MACADDR8
     }
@@ -20,7 +26,7 @@ impl crate::ToSql for macaddr::MacAddr8 {
 }
 
 #[cfg_attr(docsrs, doc(cfg(feature = "net")))]
-impl crate::FromSql for macaddr::MacAddr8 {
+impl crate::FromSql for MacAddr8 {
     /*
      * https://github.com/postgres/postgres/blob/REL_12_0/src/backend/utils/adt/mac8.c#L104
      */
@@ -46,18 +52,18 @@ impl crate::FromSql for macaddr::MacAddr8 {
 }
 
 #[cfg_attr(docsrs, doc(cfg(feature = "net")))]
-impl crate::entity::Simple for macaddr::MacAddr8 {}
+impl crate::entity::Simple for MacAddr8 {}
 
 #[cfg(test)]
 mod test {
     #![allow(non_snake_case)]
 
-    static MAC: macaddr::MacAddr8 =
-        macaddr::MacAddr8::new(0x08, 0x00, 0x2b, 0x01, 0x02, 0x03, 0x04, 0x05);
+    static MAC: crate::MacAddr8 =
+        crate::MacAddr8::new(0x08, 0x00, 0x2b, 0x01, 0x02, 0x03, 0x04, 0x05);
 
     crate::sql_test!(
         macaddr8,
-        macaddr::MacAddr8,
+        crate::MacAddr8,
         [
             (
                 "'08:00:2b:01:02:03:04:05'",