+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - `pg16` feature with new connection options;
+- Foreign types alias;
 
 ### Changed
 - `from_sql::read_*` and `from_sql::not_null` public;